Shrun

Description

This module is the entry point to the Shrun library used by the Shrun executable.

Documentation

data ShellT env (m :: Type -> Type) a

ShellT is the main application type that runs shell commands.

runShellT :: forall m env a. ShellT env m a -> env -> m a

Runs a ShellT with the given env.

shrun

Arguments

:: forall m env notifyEnv. (HasAnyError env, HasCallStack, HasCommands env, HasInit env, HasLogging env m, HasNotifyConfig env notifyEnv, HasTimeout env, MonadAsync m, MonadAtomic m, MonadEvaluate m, MonadHandleReader m, MonadHandleWriter m, MonadIORef m, MonadMask m, MonadNotify m, MonadPathReader m, MonadPathWriter m, MonadPosixFiles m, MonadPosixSignals m, MonadProcess m, MonadMVar m, MonadReader env m, MonadRegionLogger m, MonadThread m, MonadTime m, NotifyEnvF m ~ notifyEnv)

=> m ()
.

Entry point

data TermException

TermException is explicitly for when the current process is cancelled (SIGTERM on posix). We use a separate type so that we can distinguish it from potentially other ThreadKilleds that might be sent (e.g. bugs).

Constructors

MkTermException