Shrun.Command.Types

Description

Provides the Command wrapper for commands.

Command

data CommandP (p :: CommandPhase)

Wrapper for shell commands. Whenever the CommandIndex order is important, see CommandOrd.

Constructors

MkCommandP

Fields

index :: CommandIndex

NonNegative index for the command.

key :: Maybe Text

The key name for the command, for display purposes.

command :: Text

The shell command to run.

type CommandP1 = CommandP 'CommandPhase1

Phase1 commands.

type CommandP2 = CommandP 'CommandPhase2

Phase2 commands.

commandToProcess :: CommandP1 -> Maybe Text -> CreateProcess

Index

data CommandIndex

Numeric index for each command, for handling command graph dependencies. Conversion to/from Vertex should use indexToVertex and indexFromVertex, as CommandIndex is 1-based whereas Vertex is 0-based.

fromPositive :: Positive Int -> CommandIndex

unsafeFromInt :: HasCallStack => Int -> CommandIndex

toVertex :: CommandIndex -> Vertex

Conversion to Vertex.

fromVertex :: HasCallStack => Vertex -> CommandIndex

Conversion from Vertex.

succ :: CommandIndex -> CommandIndex

addNN :: CommandIndex -> NonNegative Int -> CommandIndex

range :: CommandIndex -> CommandIndex -> Either String (CommandIndex, NESeq CommandIndex)

Given m, n s.t. m < n, returns [m .. n]. Notice this is strictly increasing.

joinRange :: (a, NESeq a) -> NESeq a

Collapses range output back into a single NESeq.

Order

newtype CommandOrd (p :: CommandPhase)

Wraps CommandP for the purposes of ordering by index.

Constructors

MkCommandOrd (CommandP p)

Status

data CommandStatus

Represents a Command's status.

Constructors

CommandSuccess
The command ran successfully.

CommandFailure
The command failed.

CommandRunning (Maybe Pid, [Pid])
The command is running. It is associated to its own PID and child PIDs, for later cleanup.

CommandWaiting
The command is waiting to run.

Map

newtype CommandStatusMapP (p :: CommandStatusMapIndex)

Command status map.

Constructors

MkCommandStatusMapP

Fields

unCommandStatusMap :: CommandStatusMapIndexF p

type TCommandStatusMap = CommandStatusMapP 'CommandStatusMapStm

Normal status map, mutable in stm.

type CommandStatusMap = CommandStatusMapP 'CommandStatusMapPure

Pure status map i.e. after the stm map has been read.

readCommandStatus :: (HasCallStack, MonadAtomic m) => TCommandStatusMap -> m CommandStatusMap

Reads a map of TVars into a pure map via a single STM transaction.

Misc

data CommandPhase

Constructors

CommandPhase1

CommandPhase2

type Vertex = Int

Type for Command graph vertex, for usage with fgl.

type LVertex a = (Vertex, a)

Labeled Vertex.