Shrun.Command.Types.Internal

Description

Internal module. Take care as usage can violate invariants.

Index

newtype CommandIndex

Numeric index for each command, for handling command graph dependencies. Conversion to/from Vertex should use indexToVertex and indexFromVertex, as CommandIndex is 1-based whereas Vertex is 0-based.

Constructors

MkCommandIndex

Fields

unCommandIndex :: Positive Int

fromPositive :: Positive Int -> CommandIndex

unsafeFromInt :: HasCallStack => Int -> CommandIndex

succ :: CommandIndex -> CommandIndex

addNN :: CommandIndex -> NonNegative Int -> CommandIndex

range :: CommandIndex -> CommandIndex -> Either String (CommandIndex, NESeq CommandIndex)

Given m, n s.t. m < n, returns [m .. n]. Notice this is strictly increasing.

joinRange :: (a, NESeq a) -> NESeq a

Collapses range output back into a single NESeq.

Vertex

type Vertex = Int

Type for Command graph vertex, for usage with fgl.

type LVertex a = (Vertex, a)

Labeled Vertex.

toVertex :: CommandIndex -> Vertex

Conversion to Vertex.

fromVertex :: HasCallStack => Vertex -> CommandIndex

Conversion from Vertex.