Shrun.Configuration.Args

Description

Provides functionality for parsing command line arguments.

Since: 0.5

Documentation

data Args

Type for parsing command line args.

Since: 0.1

Constructors

MkArgs

Fields

configPath :: !(Maybe FilePath)

Optional config file.

Since: 0.5

noConfig :: Bool

Ignores toml config file.

Since: 0.5

timeout :: !(Maybe Timeout)

Timeout.

Since: 0.1

disableLogging :: !(Maybe Bool)

Global option for logging. If it is true then all logging is disabled.

Since: 0.1

cmdLogging :: !(Maybe CmdLogging)

Whether to log commands.

Since: 0.1

cmdDisplay :: !(Maybe CmdDisplay)

Whether to display command by (key) name or command.

Since: 0.1

stripControl :: !(Maybe StripControl)

Determines to what extent we should remove control characters from logs.

Since: 0.3

cmdNameTrunc :: !(Maybe (Truncation 'TCmdName))

The max number of command characters to display in the logs.

Since: 0.1

cmdLineTrunc :: !(Maybe LineTruncation)

The max number of line characters to display in the logs.

Since: 0.1

fileLogging :: !(Maybe FilePathDefault)

Optional path to log file.

Since: 0.1

fileLogStripControl :: !(Maybe StripControl)

stripControl for the log file.

Since: 0.5

commands :: !(NonEmptySeq Text)

List of commands.

Since: 0.1

defaultArgs :: NonEmptySeq Text -> Args

Default configuration.

Since: 0.1

parserInfoArgs :: ParserInfo Args

ParserInfo type for parsing Args.

Since: 0.1