Shrun.Configuration.Args.Parsing.Utils

Simple help

mkHelp :: forall (f :: Type -> Type) a. String -> Mod f a

mkHelpNoLine :: forall (f :: Type -> Type) a. String -> Mod f a

Disabled parser

withDisabledParser

Arguments

:: ReadM a
Reader for a.

-> [Mod OptionFields (WithDisabled a)]
Modifier list e.g. option name, help text.

-> (Bool, [String])
Metavar string.

-> Parser (WithDisabled a)

Constructs a parser for (WithDisabled a).

mWithDisabledParser

Arguments

:: ReadM a
Reader for a.

-> [Mod OptionFields (WithDisabled a)]
Modifier list e.g. option name, help text.

-> (Bool, [String])
Metavar string.

-> Parser (Maybe (WithDisabled a))

Constructs a parser for (Maybe (WithDisabled a)).

withDisabledParserNoOpts

Arguments

:: ReadM a
Reader for a.

-> [Mod OptionFields (WithDisabled a)]
Modifier list e.g. option name, help text.

-> Parser (WithDisabled a)

Constructs a parser for (Maybe (WithDisabled a)).

Switch parser

switchParser :: (Bool -> a) -> String -> String -> Parser (Maybe a)

switchParserNoLine :: (Bool -> a) -> String -> String -> Parser (Maybe a)

switchParserOpts :: Mod OptionFields Bool -> (Bool -> a) -> String -> String -> Parser (Maybe a)

Misc

autoStripUnderscores :: Read a => ReadM a

Reads Text, strips underscores, then uses the Read class. This is essentially auto but removes underscores. This is used for nicer numeric values e.g. allowing parsing "1_000_000" as a Num.

itemize :: NESeq String -> Mod OptionFields a

Make an itemized list e.g.

  itemize [intro, l1, l2, l2]

  ==> intro

      - l1
      - l2
      - l3

itemizeNoLine :: NESeq String -> Mod OptionFields a

itemize that does not append a trailing newline. Useful for the last option in a group, as groups already start a newline.

itemizeHelper :: NESeq String -> Chunk DocOA

toChunk :: a -> Chunk a

toMDoc :: String -> Maybe DocOA