Shrun.Configuration.Data.CommandLogging

Types

Buffer length

newtype BufferLength

Constructors

MkBufferLength Int

parseBufferLength :: MonadFail m => m Natural -> m BufferLength

Buffer timeout

newtype BufferTimeout

Constructors

MkBufferTimeout Timeout

parseBufferTimeout :: (Alternative f, MonadFail f) => f Natural -> f Text -> f BufferTimeout

Report read errors

newtype ReportReadErrorsSwitch

Switch for logging read errors

Constructors

MkReportReadErrorsSwitch Bool

Main

data CommandLoggingP (p :: ConfigPhase)

Holds config related to (console and file) command logging.

Constructors

MkCommandLoggingP

Fields

bufferLength :: ConfigPhaseF p BufferLength

Max log length held by the buffer for the ReadBlockLineBuffer ReadStrategy.

bufferTimeout :: ConfigPhaseF p BufferTimeout

Max time the buffer will hold a log before flushing it, for the ReadBlockLineBuffer ReadStrategy.

pollInterval :: ConfigPhaseF p PollInterval

How often to poll commands for logs, in microseconds.

readSize :: ConfigPhaseF p ReadSize

Determines the max log size we read from commands in one go. Note this is not on commandLogging or fileLogging since it affects both.

readStrategy :: ConfigPhaseF p ReadStrategy

Reading strategy.

reportReadErrors :: SwitchF p ReportReadErrorsSwitch

Determines if we should log read errors.

type CommandLoggingArgs = CommandLoggingP 'ConfigPhaseArgs

type CommandLoggingToml = CommandLoggingP 'ConfigPhaseToml

type CommandLoggingMerged = CommandLoggingP 'ConfigPhaseMerged

type CommandLoggingEnv = CommandLoggingP 'ConfigPhaseEnv

Functions

mergeCommandLogging :: (HasCallStack, MonadThrow m) => Bool -> Bool -> NESeq CommandP1 -> CommandLoggingArgs -> Maybe CommandLoggingToml -> m CommandLoggingMerged

Merges args and toml configs.

toEnv :: CommandLoggingMerged -> CommandLoggingEnv

Creates env version from merged. Requires commands because we pick the read strategy based on the number of commands.

defaultCommandLoggingMerged :: Bool -> Bool -> NESeq CommandP1 -> CommandLoggingP 'ConfigPhaseMerged

Exceptions

data ReadStrategyException

Constructors

MkReadStrategyException