Shrun.Configuration.Data.CommandLogging.ReadStrategy

Documentation

data ReadStrategy

Different read strategies for simplicity vs. potential prettier formatting.

Constructors

ReadBlock
Reads N bytes at a time.

ReadBlockLineBuffer
Reads N bytes at a time, but attempts to distinguish "complete" (newline terminated) vs. "partial" (anything else) reads. We do this to make the file log output prettier.

parseReadStrategy :: MonadFail m => m Text -> m ReadStrategy

Parses ReadStrategy.

readStrategyMeta :: IsString a => (Bool, [a])

Available ReadStrategy strings.

defaultReadStrategy :: Bool -> Bool -> NESeq CommandP1 -> ReadStrategy

readBlockLineBufferNotAllowed :: Bool -> Bool -> NESeq CommandP1 -> Bool