Shrun.Configuration.Data.CommonLogging.KeyHideSwitch

Documentation

data KeyHideSwitch

Type for determining if we use the command's key for display, rather than the key itself.

Constructors

KeyHideOff
Display the command's key, if it exists, rather than the key itself.

KeyHideOn
Display the command itself, not the key.