Shrun.Configuration.Data.Core

Documentation

data CoreConfigP p

Holds core configuration data.

Constructors

MkCoreConfigP

Fields

timeout :: ConfigPhaseMaybeF p Timeout

Timeout.

init :: ConfigPhaseMaybeF p Text

Shell logic to run before each command.

keyHide :: ConfigPhaseF p KeyHide

Whether to display command by (key) name or command.

pollInterval :: ConfigPhaseF p PollInterval

How often to poll commands for logs, in microseconds.

cmdLogSize :: ConfigPhaseF p (Bytes B Natural)

Determines the max log size we read from commands in one go. Note this is not on cmdLogging or fileLogging since it affects both.

timerFormat :: ConfigPhaseF p TimerFormat

How to format the timer.

cmdNameTrunc :: ConfigPhaseMaybeF p (Truncation TCmdName)

The max number of command characters to display in the logs.

cmdLogging :: ArgsOnlyDetF p (CmdLoggingP p)

Command log config.

fileLogging :: ArgsOnlyDetF p (FileLoggingP p)

File log config.

notify :: ArgsOnlyDetF p (NotifyP p)

Notify config.

type CoreConfigArgs = CoreConfigP ConfigPhaseArgs

type CoreConfigToml = CoreConfigP ConfigPhaseToml

type CoreConfigMerged = CoreConfigP ConfigPhaseMerged