Shrun.Configuration.Data.FileLogging

Documentation

data FileLogInitP (p :: ConfigPhase)

Initial file log params, for usage before we create the final Env.

Constructors

MkFileLogInitP

Fields

path :: FileLogPathF p

Optional path to log file.

sizeMode :: ConfigPhaseF p FileSizeMode

Threshold for when we should warn about the log file size.

data FileLogOpened

Params after we have opened the file for logging.

Constructors

MkFileLogOpened

Fields

handle :: ~LockedHandleW

File handle.

path :: OsPath

File path.

queue :: ~(TBQueue FileLog)

File log queue.

data FileLoggingP (p :: ConfigPhase)

Holds file logging config.

Constructors

MkFileLoggingP

Fields

file :: FileLogFileF p

File-related params.

commandNameTrunc :: ConfigPhaseDisabledMaybeF p (Truncation 'TruncCommandName)

The max number of command characters to display in the file logs.

deleteOnSuccess :: SwitchF p DeleteOnSuccessSwitch

If active, deletes the log file upon success.

lineTrunc :: LineTruncF p

Determines to what extent we should remove control characters from file logs.

mode :: ConfigPhaseF p FileMode

Mode to use with the file log.

multi :: FileLogMultiF p

Switch for multi logs. At the env stage, this is a counter for the file name index.

stripControl :: ConfigPhaseF p FileLogStripControl

Strip control

type FileLoggingArgs = FileLoggingP 'ConfigPhaseArgs

type FileLoggingToml = FileLoggingP 'ConfigPhaseToml

type FileLoggingMerged = FileLoggingP 'ConfigPhaseMerged

type FileLoggingEnv = FileLoggingP 'ConfigPhaseEnv

newtype DeleteOnSuccessSwitch

Switch for deleting the log file upon success.

Constructors

MkDeleteOnSuccessSwitch Bool

newtype FileLogMultiSwitch

Switch for logging to multiple files.

Constructors

MkFileLogMultiSwitch Bool

mergeFileLogging :: (HasCallStack, MonadCatch m, MonadIORef m, MonadTerminal m) => NESeq CommandP1 -> IORef DetectResult -> FileLoggingArgs -> Maybe FileLoggingToml -> m (Maybe FileLoggingMerged)

Merges args and toml configs.

withFileLoggingEnv :: (HasCallStack, MonadAtomic m, MonadFileWriter m, MonadHandleWriter m, MonadMask m, MonadPathReader m, MonadPathWriter m, MonadPosixFiles m, MonadTerminal m) => Maybe FileLoggingMerged -> (Maybe FileLoggingEnv -> m a) -> m a

Given merged FileLogging config, constructs a FileLoggingEnv and calls the continuation.

Misc

createMultiLogFile :: (HasCallStack, MonadAtomic m, MonadHandleWriter m, MonadPathReader m, MonadPosixFiles m, MonadThrow m) => TVar Word16 -> FileMode -> OsPath -> m OsPath

Like createLogFile, but for multi logs, which has different semantics.

While we allow overwriting / appending to existing files, what we absolutely do not want to allow is two shrun commands actively logging to the same file. This would mean logs get jumbled up.

Therefore, if we encounter an extant file, we use the TVar counter to append a number. This index is unique across all shrun commands, hence we will not have interference.