Shrun.Configuration.Data.Graph

Args

data EdgeArgs

CLI command graph. The default instance is an "edgeless graph", in the sense that all commands are root nodes without any edges, hence normal behavior.

Constructors

EdgeArgsSequential EdgeSequential
Sequential i.e. a linear graph of success edges.

EdgeArgsList Edges
Explicit edges.

newtype Edges

Dependency edges are supplied by the user on the CLI.

Constructors

MkEdges

Fields

unEdges :: Seq Edge

sortEdges :: Edges -> Edges

type Edge = (CommandIndex, CommandIndex, EdgeLabel)

An edge between two indices.

data EdgeSequential

Sequential options.

Constructors

EdgeSequentialAnd
Sequential and-edges.

EdgeSequentialOr
Sequential or-edges.

EdgeSequentialAny
Sequential any-edges.

data EdgeLabel

Types of edges.

Constructors

EdgeAnd
cmd1 & cmd2 runs cmd2 iff cmd1 succeeds.

EdgeOr
cmd1 | cmd2 runs cmd2 iff cmd1 fails.

EdgeAny
cmd1 ; cmd2 runs cmd2 iff cmd1 finishes with any status.

displayEdgeLabel :: IsString s => EdgeLabel -> s

Graph

data CommandGraph

Command dependency graph. Morally, Vertex == CommandIndex.

Constructors

MkCommandGraph

Fields

graph :: Gr CommandP1 EdgeLabel

Underlying graph.

roots :: NESeq Vertex

Root commands i.e. have no dependencies.

type Vertex = Int

Type for Command graph vertex, for usage with fgl.

type LVertex a = (Vertex, a)

Labeled Vertex.

Creation

mkGraph :: (HasCallStack, MonadThrow m) => EdgeArgs -> NESeq CommandP1 -> m CommandGraph

Creates a command dependency graph from list of dependencies and typed commands.

mkEdgelessGraph :: NESeq CommandP1 -> CommandGraph

Creates a trivial command dep graph where each command is a disconnected vertex, hence root. This exists to avoid the monad in mkGraph, as some uses want purity (defaultConfig...). We should have mkGraph mempty == mkEdgelessGraph.

Functions

labVertices :: CommandGraph -> [LVertex CommandP1]

Retrieves all labeled vertices.

labVertex :: (HasCallStack, MonadEvaluate m) => CommandGraph -> Vertex -> m (LVertex CommandP1)

Finds a labeled vertex. Can fail.

labInVertices :: CommandGraph -> Vertex -> [(Vertex, EdgeLabel)]

Given a vertex d, returns all (s, l) s.t. there exists an l-edge s -> d.

outVertices :: CommandGraph -> Vertex -> [Vertex]

Given a vertex s, returns all d s.t. there exists an edge s -> d.

vertices :: CommandGraph -> [Vertex]

Retrieves all vertices.

Context

context :: (HasCallStack, MonadEvaluate m) => CommandGraph -> Vertex -> m (Context CommandP1 EdgeLabel)

Given a vertex, retrieves its context.

ctxLabVertex :: Context a b -> LVertex a

labVertex that operates on the context.

ctxOutVertices :: Context a b -> [Vertex]

outVertices that operates on the context.