Shrun.Configuration.Data.Notify

NotifyInit

data NotifyActionsInit (p :: ConfigPhase)

Initial notify action config.

Constructors

MkNotifyActionsInit

Fields

complete :: NotifyActionCompleteF p

start :: SwitchF p NotifyActionStartSwitch

NotifyActive

data NotifyActionsActive

Holds active notif actions. This is morally These, specialized to our notify action use case.

Constructors

NotifyActionsActiveComplete NotifyActionComplete
Complete actions active.

NotifyActionsActiveStart
Start actions active.

NotifyActionsActiveAll NotifyActionComplete
All actions active.

_NotifyActionsActiveCompleteAny :: AffineFold NotifyActionsActive NotifyActionComplete

_NotifyActionsActiveStartAny :: AffineFold NotifyActionsActive ()

Notify

data NotifyP (p :: ConfigPhase) r

Holds notification config. We have an invariant that if the notify config exists (i.e. is Just), then at least one of actionComplete, actionStart should be Just/True.

Constructors

MkNotifyP

Fields

actions :: NotifyActionsF p

Notify actions.

system :: NotifySystemF p r

The notification system to use.

timeout :: ConfigPhaseF p NotifyTimeout

when to timeout successful notifications.

type NotifyArgs = NotifyP 'ConfigPhaseArgs

type NotifyToml = NotifyP 'ConfigPhaseToml

type NotifyMerged = NotifyP 'ConfigPhaseMerged

type NotificationEnv = NotifyP 'ConfigPhaseEnv

mergeNotifications :: NotifyArgs r -> Maybe (NotifyToml r) -> Maybe (NotifyMerged r)

Merges args and toml configs.

toEnv :: (HasCallStack, MonadNotify m, MonadThrow m, NotifyEnvF m ~ r) => NotifyMerged r -> m (NotifyP 'ConfigPhaseEnv r)