Shrun.Configuration.Data.Notify.Action

Description

Provides type for notifications.

Start

newtype NotifyActionStartSwitch

Type for determining if we use the command's key for display, rather than the key itself.

Constructors

MkNotifyActionStartSwitch Bool

Complete

data NotifyActionComplete

Determines for which complete actions we should send notifications.

Constructors

NotifyActionCompleteFinal
Send a notification after all commands are completed.

NotifyActionCompleteCommand
Send notifications when each command completes.

NotifyActionCompleteAll
NotifyActionCompleteFinal and NotifyActionCompleteCommand.

parseNotifyAction :: MonadFail m => Text -> m NotifyActionComplete

Parses NotifyActionComplete.

notifyActionMeta :: IsString a => (Bool, [a])

Available NotifyActionComplete strings.