Shrun.Configuration.Data.StripControl

Documentation

data StripControl (t :: StripControlType)

Determines how we should treat control characters encountered in logs.

Constructors

StripControlSmart
"Intelligently" strip control characters e.g. colors are fine, ones that affect the cursor should be removed.

StripControlNone
Do not strip any control characters.

StripControlAll
Strip all control characters.

parseStripControl :: forall m (t :: StripControlType). MonadFail m => m Text -> m (StripControl t)

type ConsoleLogStripControl = StripControl 'StripControlConsoleLog

type FileLogStripControl = StripControl 'StripControlFileLog