Shrun.Configuration.Data.Truncation

Documentation

data TruncRegion

The different regions to apply truncation rules.

Constructors

TruncCommandName
Apply truncation rules to commands/key names.

TruncLine
Apply truncation rules to command log entire lines.

newtype Truncation (a :: TruncRegion)

The maximum number of command characters to display in the logs.

Constructors

MkTruncation

Fields

unTruncation :: Int

parseTruncation :: forall m (r :: TruncRegion). MonadFail m => m Natural -> m (Truncation r)

data LineTruncation

Determines command log line truncation behavior. We need a separate type from Truncation to add a third option, to detect the terminal size automatically.

Constructors

Undetected (Truncation 'TruncLine)

Detected

parseLineTruncation :: (Alternative m, MonadFail m) => m Natural -> m Text -> m LineTruncation

Misc

decodeCommandNameTrunc :: Decoder (Maybe (WithDisabled (Truncation 'TruncCommandName)))

decodeLineTrunc :: Decoder (Maybe (WithDisabled LineTruncation))

data DetectResult

Terminal width detect result. We save this so we only look it up at most once (i.e. both --console-log-line-trunc and --file-log-line-trunc are on).

Constructors

DetectNotRun

DetectFailed

DetectSucceeded Int

mergeLineTrunc

Arguments

:: (HasCallStack, MonadCatch m, MonadIORef m, MonadTerminal m)

=> Bool
If true, defaults to detect.

-> IORef DetectResult
Detect result cache, so we only look it up at most once.

-> Maybe (WithDisabled LineTruncation)
Args config.

-> Maybe (WithDisabled LineTruncation)
Toml config.

-> m (Maybe (Truncation 'TruncLine))

Merges line truncation.

lineTruncMeta :: IsString a => (Bool, [a])