Shrun.Configuration.Default

Documentation

class Default a where

For types with a default value. In general, instances should be "simple" i.e. no instances for aggregate TTG types (e.g. FileLogging) except for CLI args, as complexity jumps quickly.

For the most part, types should only have instances for one of Default xor (SemigroupMonoid or Alternative) i.e. if types do/ support some sort of algebra, the latter is preferred. Default exists when no algebra is sensible.

Methods

def :: a

fromMaybe :: Default a => Maybe a -> a

(<.>) :: Default a => Maybe a -> Maybe a -> a infixr 6

Like (<>?) except we extract a result via fromDefault.