Shrun.Configuration.Env

Description

Module that provides env types and requisite typeclasses, along with parsing functionality.

Since: 0.1

"HasX" style typeclasses

class HasCommands env where

The commands themselves.

Since: 0.1

Methods

getCommands :: env -> NonEmptySeq Command

Since: 0.1

class HasCompletedCmds env where

Determines command line truncation behavior.

Since: 0.1

Methods

getCompletedCmds :: env -> TVar (Seq Command)

Since: 0.1

class HasLogging env where

Holds logging configuration.

Since: 0.3

Methods

getCmdDisplay :: env -> CmdDisplay

Determines how to display command names.

Since: 0.1

getCmdLineTrunc :: env -> Maybe (Truncation 'TCmdLine)

Determines command line truncation behavior.

Since: 0.1

getCmdLogging :: env -> CmdLogging

Determines if we should log commands' output to the console.

Since: 0.1

getCmdNameTrunc :: env -> Maybe (Truncation 'TCmdName)

Determines command name truncation behavior.

Since: 0.1

getFileLogging :: env -> Maybe (FilePath, LogTextQueue)

File logging, if any.

Since: 0.3

getFileLogStripControl :: env -> StripControl

Determines control character behavior for file logs.

Since: 0.5

getDisableLogging :: env -> Bool

Determines if logging is enabled globally.

Since: 0.1

getStripControl :: env -> StripControl

Determines control character behavior.

Since: 0.3

class HasTimeout env where

Timeout, if any.

Since: 0.1

Methods

getTimeout :: env -> Maybe Timeout

Since: 0.1

Types

data Env

The main Env type used by Shrun. Intended to be used with MonadReader.

Since: 0.1

Constructors

MkEnv

Fields

timeout :: !(Maybe Timeout)

Timeout.

Since: 0.1

fileLogging :: !(Maybe (Tuple2 FilePath LogTextQueue))

Optional file logging. If enabled, holds the path to the file and the log queue.

Since: 0.1

fileLogStripControl :: !StripControl

Determines to what extent we should remove control characters from file logs.

Since: 0.3

cmdLogging :: !CmdLogging

Whether to log commands.

Since: 0.1

cmdDisplay :: !CmdDisplay

Whether to display the command (key) names or the commands themselves.

Since: 0.1

cmdNameTrunc :: !(Maybe (Truncation 'TCmdName))

The max number of command characters to display in the logs.

Since: 0.1

cmdLineTrunc :: !(Maybe (Truncation 'TCmdLine))

The max number of line characters to display in the logs.

Since: 0.1

stripControl :: !StripControl

Determines to what extent we should remove control characters from logs.

Since: 0.3

completedCmds :: !(TVar (Seq Command))

Holds a sequence of commands that have completed. Used so we can determine which commands have not completed if we time out.

Since: 0.1

disableLogging :: !Bool

Overarching option for logging. If it is false then all logging is disabled.

Since: 0.1

commands :: !(NonEmptySeq Command)

The commands to run.

Since: 0.1

data CmdDisplay

Type for determining if we use the command's key for display, rather than the key itself.

Constructors

ShowKey
Display the command's key, if it exists, rather than the key itself.

Since: 0.1

HideKey
Display the command itself, not the key.

Since: 0.5

data CmdLogging

Type for determining if we stream commands' logs.

Since: 0.1

Constructors

Disabled
No logging of sub-commands.

Since: 0.1

Enabled
Logging of sub-commands

Since: 0.1

newtype Truncation a

The maximum number of command characters to display in the logs.

Since: 0.1

Constructors

MkTruncation

Fields

unTruncation :: Natural

Since: 0.1

data TruncRegion

The different regions to apply truncation rules.

Since: 0.1

Constructors

TCmdName
Apply truncation rules to commands/key names.

Since: 0.1

TCmdLine
Apply truncation rules to command log entire lines.

Since: 0.1

data StripControl

Determines how we should treat control characters encountered in logs.

Since: 0.3

Constructors

StripControlSmart
"Intelligently" strip control characters e.g. colors are fine, ones that affect the cursor should be removed.

Since: 0.3

StripControlNone
Do not strip any control characters.

Since: 0.3

StripControlAll
Strip all control characters.

Since: 0.3

newtype TomlError

Since: 0.5

Constructors

MkTomlError

Fields

unTomlError :: TOMLError

Since: 0.5

Functions

makeEnv :: (FileSystemReader m, MonadIO m, Mutable m, Terminal m) => m Env

Returns the Env based on the CLI args and TOML config file. CLI args take precedence over config file.

Since: 0.1