Shrun.Configuration.Env.Types

Description

Provides types and typeclasses for our environment.

"HasX" style typeclasses

class HasCommands env where

The commands themselves.

Methods

getCleanup :: env -> Maybe CommandCleanup

Retrieves the cleanup functions, if they exist.

getCommandDepGraph :: env -> CommandGraph

Retrieves full command graph.

getCommandStatusMap :: env -> TCommandStatusMap

Retrieves commands and their statuses.

data CommandCleanup

Constructors

MkCommandCleanup

Fields

findPidsExe :: FilePath

killPidsExe :: FilePath

updateCommandStatus :: (HasCallStack, HasCommands env, MonadAtomic m, MonadReader env m, MonadThrow m) => CommandP1 -> CommandStatus -> m ()

Prepends a completed command.

getReadCommandStatus :: (HasCallStack, HasCommands env, MonadAtomic m, MonadReader env m) => m CommandStatusMap

Retrieves the entire status map in a single STM transaction.

class HasCommandLogging env where

Methods

getCommandLogging :: env -> CommandLoggingEnv

class HasCommonLogging env where

Methods

getCommonLogging :: env -> CommonLoggingEnv

class HasConsoleLogging env r where

Methods

getConsoleLogging :: env -> (ConsoleLoggingEnv, TBQueue (LogRegion r), IORef (Maybe r))

class HasFileLogging env where

Methods

getFileLogging :: env -> Maybe FileLoggingEnv

class HasTimeout env where

Timeout, if any.

Methods

getTimeout :: env -> WithDisabled Timeout

getHasTimedOut :: env -> TVar Bool

setTimedOut :: (HasTimeout env, MonadAtomic m, MonadReader env m) => m ()

Sets timedout to true.

whenTimedOut :: (HasTimeout env, MonadAtomic m, MonadReader env m) => m () -> m ()

Run the action when shrun has timed out.

class HasInit env where

Init, if any.

Methods

getInit :: env -> Maybe Text

class HasAnyError env where

Methods

getAnyError :: env -> TVar Bool

Retrieves the anyError flag.

setAnyErrorTrue :: (HasAnyError env, HasCallStack, MonadAtomic m, MonadReader env m) => m ()

Set anyError to True.

class HasNotifyConfig env r where

Class for retrieving the notify config.

Methods

getNotifyConfig :: env -> Maybe (NotificationEnv r)

Retrieves the notify config.

Aggregate

type HasLogging env (m :: Type -> Type) = (HasCommandLogging env, HasCommonLogging env, HasConsoleLogging env (Region m), HasFileLogging env)

Alias for all logging config.

Types

data Env notifyEnv logRegion

The main Env type used by Shrun.

Constructors

MkEnv

Fields

anyError :: TVar Bool

Holds the anyError flag, signaling if any command exited with an error.

commandCleanup :: Maybe CommandCleanup

Functions to clean up running commands.

commands :: NESeq CommandP1

Holds notification environment. | Commands

commandGraph :: CommandGraph

Command graph.

commandStatusMap :: TCommandStatusMap

Map from CommandIndex to Command and its status. Used for determining e.g. which commands have completed failed not run.

The statuses are TVars since they are mutable, though the map itself can be pure since its structure is fixed at initialization. In fact, we could probably swap TVar for IORef since we only update the status from a single thread (each command has its own thread).

config :: CoreConfigP 'ConfigPhaseEnv notifyEnv

Core config.

consoleLogQueue :: ~(TBQueue (LogRegion logRegion))

Console log queue.

hasTimedOut :: TVar Bool

timerRegion :: IORef (Maybe logRegion)

Timer region. It's an IORef only because it is not initialized on startup. Once it is set it is no longer mutated.

whenDebug :: (HasCommonLogging env, MonadReader env m) => m () -> m ()

Run the action when the debug flag is active.