Shrun.Configuration.Toml

Documentation

data Toml notifyEnv

Holds toml config.

Constructors

MkToml

Fields

coreConfig :: CoreConfigToml notifyEnv

Core config.

legend :: Maybe (Seq KeyVal)

Legend.

mergeTomls :: Seq (Toml notifyEnv) -> Toml notifyEnv

Note that our Semigroup is not commutative, hence the order matters. In particular, mconcat is safe because it is foldr, hence respects the input order.