Shrun.Data.NonEmptySeq

Description

Provides the NonEmptySeq type.

Since: 0.1

Type

data NonEmptySeq a

Represents a non-empty sequence. This is useful for when we want a non-empty, finite list.

Since: 0.1

Constructors

a :|^ (Seq a) infixr 5

Construction

singleton :: a -> NonEmptySeq a

Creates a singleton NonEmptySeq.

Since: 0.1

fromList :: List a -> Maybe (NonEmptySeq a)

Returns Just NonEmptySeq if the list is non-empty. Otherwise returns Nothing.

Since: 0.5

fromNonEmpty :: NonEmpty a -> NonEmptySeq a

NonEmptySeq from NonEmpty. Unsafe in the sense that it does not terminate for infinite NonEmpty.

Examples

Expand

>>> fromNonEmpty ( 1 :| [3,4])
1 :|^ fromList [3,4]

Since: 0.1

Elimination

toSeq :: NonEmptySeq a -> Seq a

Exposes the underlying sequence.

Examples

Expand

>>> toSeq (1 :|^ Exts.fromList [2,3,4])
fromList [1,2,3,4]

Since: 0.1

toList :: NonEmptySeq a -> List a

NonEmptySeq to list.

Examples

Expand

>>> toList (1 :|^ Exts.fromList [2,3,4])
[1,2,3,4]

Since: 0.1

Unsafe

unsafeFromList :: HasCallStack => List a -> NonEmptySeq a

Even more unsafe than fromNonEmpty; will throw an error on an empty list.

Examples

Expand

>>> unsafeFromList [1,2,3]
1 :|^ fromList [2,3]

Since: 0.1