Shrun.Data.StripControl

Documentation

data StripControl

Determines how we should treat control characters encountered in logs.

Constructors

StripControlSmart
"Intelligently" strip control characters e.g. colors are fine, ones that affect the cursor should be removed.

StripControlNone
Do not strip any control characters.

StripControlAll
Strip all control characters.

parseStripControl :: MonadFail m => m Text -> m StripControl