Shrun.Data.Text

Documentation

newtype UnlinedText

Text after it has had all lines separated into different texts. We introduce a newtype for clarity. The idea is that when we read arbitrary text from a handle, this type serves as a witness that we have indeed split the string along newlines. Then, in the normal case, we log each line separately.

In exceptional cases (e.g. command names), we may choose to combine the list back into a single text, according to some logic.

The constructor UnsafeUnlinedText should only be used when we know the text has no newlines and performance means a branch is undesirable (e.g. streaming). If there is no performance impact, consider unsafeUnlinedText instead.

Constructors

UnsafeUnlinedText

Fields

unUnlinedText :: Text

Creation

fromText :: Text -> [UnlinedText]

Creates a list of UnlinedText.

fromTextReplace :: Text -> UnlinedText

Creates a single UnlinedText by replacing newlines with whitespace.

unsafeFromTextNE :: HasCallStack => Text -> NonEmpty UnlinedText

Requires input text to be non-empty to be safe.

unsafeUnlinedText :: HasCallStack => Text -> UnlinedText

Unsafe creation that throws error when the text contains newline(s).

Elimination

toText :: [UnlinedText] -> Text

Functions

length :: UnlinedText -> Int

concat :: [UnlinedText] -> UnlinedText

Concats via toText.

intercalate :: UnlinedText -> [UnlinedText] -> UnlinedText

isWhitespace :: UnlinedText -> Bool

reallyUnsafeMap :: (Text -> Text) -> UnlinedText -> UnlinedText

Lifts a Text function to UnlinedText. Very unsafe in that we do not check for errors i.e. if the parameter function introduces any newlines, then this will silently succeed. This exists for performance.