Shrun.Data.Truncation

Documentation

data TruncRegion

The different regions to apply truncation rules.

Constructors

TCmdName
Apply truncation rules to commands/key names.

TCmdLine
Apply truncation rules to command log entire lines.

newtype Truncation a

The maximum number of command characters to display in the logs.

Constructors

MkTruncation

Fields

unTruncation :: Natural

parseTruncation :: MonadFail m => m Natural -> m (Truncation r)

data LineTruncation

Determines command log line truncation behavior. We need a separate type from Truncation to add a third option, to detect the terminal size automatically.

Constructors

Undetected (Truncation TCmdLine)

Detected

parseLineTruncation :: (Alternative m, MonadFail m) => m Natural -> m Text -> m LineTruncation