Shrun.Effects.Process

Description

Provides the Process typeclass.

Since: 0.5

Documentation

class (Monad m, RegionLogger m, Region m ~ ConsoleRegion) => Process m where

Effect for launching a process. Provides three functions of varying logging behavior.

Since: 0.5

Methods

tryCmd :: Command -> m (Maybe Stderr)

Runs the command and returns an error message on failure.

Since: 0.5

tryCmdStream :: Command -> m (Maybe Stderr)

We stream the commands' output like tryStreamRegion except we do not create a console region. This function is intended for when we want to send command logs to a file, but do not want to stream them to the console.

Since: 0.5

tryCmdStreamRegion :: Command -> m (Maybe Stderr)

Similar to tryCmd except we attempt to stream the commands' output to a Region.

Since: 0.5

tryTimeCmd :: (Process m, Timing m) => (Command -> m (Maybe Stderr)) -> Command -> m (Either (Tuple2 RelativeTime Stderr) RelativeTime)

Runs a command with timing. Returns an Either so that the user is forced to check for failures.

Since: 0.5