Shrun.Env.Types

Description

Provides types and typeclasses for our environment.

"HasX" style typeclasses

class HasCommands env where

The commands themselves.

Methods

getCommands :: env -> NESeq CommandP1

getCompletedCmds :: env -> TVar (Seq CommandP1)

prependCompletedCommand :: (HasCallStack, HasCommands env, MonadReader env m, MonadSTM m) => CommandP1 -> m ()

Prepends a completed command.

class HasLogging env r where

Holds logging configuration.

Methods

getLogging :: env -> Logging r

Retrieves logging env.

class HasTimeout env where

Timeout, if any.

Methods

getTimeout :: env -> Maybe Timeout

class HasInit env where

Init, if any.

Methods

getInit :: env -> Maybe Text

class HasAnyError env where

Methods

getAnyError :: env -> TVar Bool

Retrieves the anyError flag.

setAnyErrorTrue :: (HasAnyError env, HasCallStack, MonadReader env m, MonadSTM m) => m ()

Set anyError to True.

class HasNotifyConfig env where

Class for retrieving the notify config.

Methods

getNotifyConfig :: env -> Maybe NotifyConfig

Retrieves the notify config.

Types

data Env

The main Env type used by Shrun.

Constructors

MkEnv

Fields

timeout :: Maybe Timeout

Timeout.

init :: Maybe Text

Shell logic to run before each command.

logging :: Logging ConsoleRegion

Logging env.

completedCmds :: TVar (Seq CommandP1)

Holds a sequence of commands that have completed. Used so we can determine which commands have not completed if we time out.

anyError :: TVar Bool

Holds the anyError flag, signaling if any command exited with an error.

notifyEnv :: Maybe NotifyEnv

Holds notification environment.

commands :: NESeq CommandP1

The commands to run.

data Logging r

Holds logging data.

Constructors

MkLogging

Fields

keyHide :: KeyHide

Whether to display the command (key) names or the commands themselves.

pollInterval :: PollInterval

How often to poll commands for logs, in microseconds.

cmdLogSize :: Bytes B Natural

The max number of command characters to display in the logs.

timerFormat :: TimerFormat

How to format the timer.

cmdNameTrunc :: Maybe (Truncation TCmdName)

Truncates command names in the logs.

cmdLog :: Maybe CmdLogging

Whether to log commands.

consoleLog :: ~(TBQueue (LogRegion r))

Console log queue.

fileLog :: Maybe FileLogging

Optional file logging. If enabled, holds the path to the file and the log queue.

data CmdLogging

Constructors

MkCmdLogging

Fields

stripControl :: StripControl

lineTrunc :: Maybe (Truncation TCmdLine)

data FileLogging

Constructors

MkFileLogging

Fields

stripControl :: StripControl

log :: ~(Tuple2 Handle (TBQueue FileLog))

data NotifyEnv

Holds notification settings.

Constructors

MkNotifyEnv

Fields

system :: !NotifySystemP2

Notification system to use.

action :: NotifyAction

Notification action.

timeout :: NotifyTimeout

Timeout to use for notifications.