Shrun.IO

Description

Provides the low-level IO functions for running shell commands.

Since: 0.1

Stdout/stderr newtypes

newtype Stdout

Newtype wrapper for stdout.

Since: 0.1

Constructors

MkStdout

Fields

getStdout :: Text

Since: 0.1

newtype Stderr

Newtype wrapper for stderr.

Since: 0.1

Constructors

MkStderr

Fields

getStderr :: Text

Since: 0.1

Timing shell programs

tryCommand :: (HasCompletedCmds env, MonadIO m, MonadReader env m) => Command -> m (Maybe Stderr)

Version of tryShExitCode with timing. On success, stdout is not returned.

Since: 0.1

tryCommandStreamRegion :: (HasCompletedCmds env, HasLogging env, MonadMask m, MonadReader env m, Mutable m, MonadUnliftIO m, RegionLogger m, Region m ~ ConsoleRegion, Timing m) => Command -> m (Maybe Stderr)

Similar to tryCommand except we attempt to stream the commands' output to a ConsoleRegion instead of the usual swallowing.

Since: 0.1

tryCommandStreamNoRegion :: (HasCompletedCmds env, HasLogging env, MonadReader env m, Mutable m, MonadUnliftIO m, RegionLogger m, Region m ~ ConsoleRegion, Timing m) => Command -> m (Maybe Stderr)

We stream the commands' output like tryCommandStreamRegion except we do not create a console region. This function is intended for when we want to send command logs to a file, but do not want to stream them to the console.

Since: 0.1