Shrun.Logging.Log

Description

This module provides convenience functions for formatting and sending logs to destinations.

Since: 0.1

High-level

putLog :: (HasLogging env, MonadReader env m, Mutable m, RegionLogger m, Timing m) => Log -> m ()

Conditionally writes a log to the console and file, depending on the HasLogging environment.

Since: 0.3

putRegionLog :: (HasLogging env, MonadReader env m, Mutable m, RegionLogger m, Timing m) => Region m -> Log -> m ()

Conditionally writes a log to the console region and file, depending on the HasLogging environment.

Since: 0.3

Low-level

maybePrintLog :: (HasLogging env, MonadReader env m) => (Text -> m ()) -> Log -> m ()

maybePrintLog fn log applies fn if the log has dest LogFile. Otherwise does nothing.

Since: 0.3

maybeSendLogToQueue :: (HasLogging env, MonadReader env m, Mutable m, Timing m) => Log -> m ()

Sends the log to the file queue as long as the dest is not LogConsole.

Since: 0.3