Shrun.Logging.MonadRegionLogger

Description

Provides functionality for logging to a specific region (i.e. for concurrent console logging).

Documentation

class Monad m => MonadRegionLogger (m :: Type -> Type) where

MonadRegionLogger is a simple typeclass for abstracting logging functions.

Associated Types

type Region (m :: Type -> Type)

The type of the region. This will be ConsoleRegion for production code and () for tests.

Methods

logGlobal :: Text -> m ()

Pushes a log to the "global" region.

logRegion :: LogMode -> Region m -> Text -> m ()

Pushes a log to the region.

withRegion :: HasCallStack => RegionLayout -> (Region m -> m a) -> m a

Runs an m a with a region.

displayRegions :: HasCallStack => m a -> m a

Displays the regions. This should wrap whatever top-level logic needs regions.