Shrun.Logging.Queue

Description

Provides the LogTextQueue type and associated functions. This is intended for concurrently writing logs to a file.

Since: 0.1

LogText

data LogText where

LogText is a textual representation of a given Log. No coloring is included, but we include the prefix (e.g. Warn) along with a timestamp.

Since: 0.1

Bundled Patterns

pattern MkLogText :: Text -> LogText
Since: 0.1

_MkLogText :: Getter LogText Text

Since: 0.5

formatFileLog :: (HasLogging env, MonadReader env m, Timing m) => Log -> m LogText

Formats a Log into a LogText. Applies prefix and timestamp.

Examples

Expand

formatFileLog $ MkLog Nothing "Running time: 2 seconds" Info Set LogBoth
"[2022-02-23 20:58:04.231933782 UTC] [Info] Running time: 2 secondsn"

formatFileLog $ MkLog (Just "cmd") "cmd: command not found" Error Set LogBoth
"[2022-02-23 20:58:04.231933782 UTC] [Error] [cmd] cmd: command not foundn"

Since: 0.1

Queue

newtype LogTextQueue

Newtype wrapper over a TBQueue.

Since: 0.1

Constructors

MkLogTextQueue

Fields

getLogTextQueue :: TBQueue LogText

Since: 0.1

_MkLogTextQueue :: Iso' LogTextQueue (TBQueue LogText)

Since: 0.5

readQueue :: Mutable m => LogTextQueue -> m (Maybe LogText)

Atomically reads from the queue. Does not retry.

Since: 0.1

writeQueue :: (HasLogging env, MonadReader env m, Mutable m, Timing m) => LogTextQueue -> Log -> m ()

Atomically writes to the queue.

Since: 0.1

flushQueue :: Mutable m => LogTextQueue -> m (List LogText)

Atomically flushes the queue's entire contents. Does not retry.

Since: 0.1