Shrun.Logging.RegionLogger

Description

Provides functionality for logging to a specific region (i.e. for concurrent console logging).

Since: 0.1

Typeclass for logging

class Monad m => RegionLogger m where

RegionLogger is a simple typeclass for abstracting logging functions.

Since: 0.1

Associated Types

type Region m

The region type for our logging. For real programs (e.g. IO), this will likely be ConsoleRegion.

Since: 0.1

Methods

logFn :: Text -> m ()

Pushes a log to the "global" region.

Since: 0.1

logModeToRegionFn :: LogMode -> Region m -> Text -> m ()

Selects a region fn based on the LogMode.

Since: 0.3

withConsoleRegion :: RegionLayout -> (Region m -> m a) -> m a

Runs an m a with a console region.

Since: 0.5