Shrun.Logging.Types

Description

This module provides the Log type and associated functions.

Since: 0.3

Types for logging

data Log

Captures the relevant information concerning a specific log (i.e. text, level, and mode).

Since: 0.1

Constructors

MkLog

Fields

cmd :: Maybe Command

Optional command that produced this log.

Since: 0.1

msg :: Text

The Text for a given log.

Since: 0.1

lvl :: LogLevel

The LogLevel for a given log.

Since: 0.1

mode :: LogMode

The LogMode for a given log.

Since: 0.1

dest :: LogDest

Where to send this log. Most logs should go to both the console and the file. For a log to actually be written to a file, dest must be either LogFile or LogBoth and file logging must be enabled globally.

Since: 0.1

data LogMode

Determines the logging behavior.

Since: 0.1

Constructors

Set
Sets the logging region to this log.

Since: 0.1

Append
Appends the log to the logging region.

Since: 0.1

Finish
Closes the logging region, finishing with the log.

Since: 0.1

data LogLevel

Determines the logging level.

Since: 0.1

Constructors

None
Since: 0.1

SubCommand
Since: 0.1

Debug
Since: 0.1

Info
Since: 0.1

InfoBlue
Since: 0.1

InfoCyan
Since: 0.1

InfoSuccess
Since: 0.1

Warn
Since: 0.1

Error
Since: 0.1

Fatal
Since: 0.1

data LogDest

Determines where the log is sent.

Since: 0.1

Constructors

LogConsole
Since: 0.1

LogFile
Since: 0.1

LogBoth
Since: 0.1

Utility functions for associate levels to colors/prefixes.

logToColor :: Log -> Color

Transforms log to a color based on its LogLevel.

Since: 0.1

logToPrefix :: Log -> Text

Transforms log to a prefix based on its LogLevel.

Since: 0.1

levelToColor :: LogLevel -> Color

Maps LogLevel to Color.

Since: 0.1

levelToPrefix :: LogLevel -> Text

Maps LogLevel to 'Prefix'.

Since: 0.1

Log Queue

newtype LogText

LogText is a textual representation of a given Log. No coloring is included, but we include the prefix (e.g. Warn) along with a timestamp.

Since: 0.1

Constructors

UnsafeLogText

Fields

unLogText :: Text

Since: 0.1

Bundled Patterns

pattern MkLogText :: Text -> LogText
Since: 0.1

_MkLogText :: Getter LogText Text

Since: 0.5

newtype LogTextQueue

Newtype wrapper over a TBQueue.

Since: 0.1

Constructors

MkLogTextQueue

Fields

getLogTextQueue :: TBQueue LogText

Since: 0.1

_MkLogTextQueue :: Iso' LogTextQueue (TBQueue LogText)

Since: 0.5