Shrun.Logging.Types.Internal

Description

Internal module for logging types.

Documentation

newtype FileLog

FileLog is a textual representation of a given log after it has been formatted. No coloring is included, but we include the prefix (e.g. Warn) along with a timestamp.

Constructors

UnsafeFileLog

Fields

unFileLog :: Text

newtype ConsoleLog

ConsoleLog is a textual representation of a given log after it has been formatted.

Constructors

UnsafeConsoleLog

Fields

unConsoleLog :: Text

data LogMode

Determines the logging behavior. This option only affects console logs; File logs are unaffected.

Constructors

LogModeSet
Sets the logging region to this log.

LogModeAppend
Appends the log to the logging region.

LogModeFinish
Closes the logging region, finishing with the log.