Shrun

Description

This module is the entry point to the Shrun library used by the Shrun executable.

Since: 0.1

Documentation

data ShellT env m a

ShellT is the main application type that runs shell commands.

Since: 0.1

runShellT :: ShellT env m a -> env -> m a

Runs a ShellT with the given env.

Since: 0.1

shrun :: (FileSystemWriter m, HasCommands env, HasCompletedCmds env, HasLogging env, HasTimeout env, MonadMask m, MonadReader env m, MonadUnliftIO m, Mutable m, Process m, Terminal m, Timing m) => m ()

Entry point

Since: 0.1