Shrun

Description

This module is the entry point to the Shrun library used by the Shrun executable.

Documentation

data ShellT env m a

ShellT is the main application type that runs shell commands.

runShellT :: ShellT env m a -> env -> m a

Runs a ShellT with the given env.

shrun

Arguments

:: (HasAnyError env, HasCommands env, HasInit env, HasLogging env (Region m), HasNotifyConfig env, HasTimeout env, MonadAsync m, MonadHandleReader m, MonadHandleWriter m, MonadIORef m, MonadNotify m, MonadTypedProcess m, MonadMask m, MonadReader env m, MonadRegionLogger m, MonadSTM m, MonadThread m, MonadTime m)

=> m ()
.

Entry point