Shrun.Configuration.Args.Parsing.Utils

Documentation

withDisabledParser

Arguments

:: Parser (Maybe a)
Main parser.

-> String
Name for this option, to be used in disabled switch name.

-> Parser (WithDisabled a)

Adds a '--no-x' switch to the parser.

withDisabledParserOpts

Arguments

:: Mod FlagFields Bool
Disabled switch options.

-> Parser (Maybe a)
Main parser

-> String
Name for this option, to be used in disabled switch name.

-> Parser (WithDisabled a)

Like withDisabledParser, except it also takes an arg for the disabled switch options.

mkHelp :: forall (f :: Type -> Type) a. String -> Mod f a

autoStripUnderscores :: Read a => ReadM a

Reads Text, strips underscores, then uses the Read class. This is essentially auto but removes underscores. This is used for nicer numeric values e.g. allowing parsing "1_000_000" as a Num.