Shrun.Configuration.Data.CommandLogging

Types

data ReportReadErrorsSwitch

Switch for logging read errors

Constructors

ReportReadErrorsOff

ReportReadErrorsOn

data CommandLoggingP (p :: ConfigPhase)

Holds config related to (console and file) command logging.

Constructors

MkCommandLoggingP

Fields

pollInterval :: ConfigPhaseF p PollInterval

How often to poll commands for logs, in microseconds.

readSize :: ConfigPhaseF p ReadSize

Determines the max log size we read from commands in one go. Note this is not on commandLogging or fileLogging since it affects both.

reportReadErrors :: SwitchF p ReportReadErrorsSwitch

Determines if we should log read errors.

type CommandLoggingArgs = CommandLoggingP 'ConfigPhaseArgs

type CommandLoggingToml = CommandLoggingP 'ConfigPhaseToml

type CommandLoggingMerged = CommandLoggingP 'ConfigPhaseMerged

type CommandLoggingEnv = CommandLoggingP 'ConfigPhaseEnv

Functions

mergeCommandLogging :: CommandLoggingArgs -> Maybe CommandLoggingToml -> CommandLoggingMerged

Merges args and toml configs.

toEnv :: CommandLoggingMerged -> CommandLoggingEnv

Creates env version from merged.

Misc

defaultMerged :: CommandLoggingMerged