Shrun.Configuration.Data.FileLogging

Documentation

data FileLoggingP p

Holds file logging config.

Constructors

MkFileLoggingP

Fields

path :: FileLogPathF p

Optional path to log file.

stripControl :: ConfigPhaseF p StripControl

Determines to what extent we should remove control characters from file logs.

mode :: ConfigPhaseF p FileMode

Mode to use with the file log.

sizeMode :: ConfigPhaseF p FileSizeMode

Threshold for when we should warn about the log file size.

type FileLoggingArgs = FileLoggingP ConfigPhaseArgs

type FileLoggingToml = FileLoggingP ConfigPhaseToml

type FileLoggingMerged = FileLoggingP ConfigPhaseMerged

mergeFileLogging :: FileLoggingArgs -> Maybe FileLoggingToml -> Maybe FileLoggingMerged

Merges args and toml configs.