Shrun.Configuration.Data.FileLogging

Documentation

data FileLogInitP (p :: ConfigPhase)

Initial file log params, for usage before we create the final Env.

Constructors

MkFileLogInitP

Fields

path :: FileLogPathF p

Optional path to log file.

mode :: ConfigPhaseF p FileMode

Mode to use with the file log.

sizeMode :: ConfigPhaseF p FileSizeMode

Threshold for when we should warn about the log file size.

data FileLogOpened

Params after we have opened the file for logging.

Constructors

MkFileLogOpened

Fields

handle :: ~Handle

File handle.

queue :: ~(TBQueue FileLog)

File log queue.

data FileLoggingP (p :: ConfigPhase)

Holds file logging config.

Constructors

MkFileLoggingP

Fields

file :: FileLogFileF p

File-related params.

commandNameTrunc :: ConfigPhaseMaybeF p (Truncation 'TruncCommandName)

The max number of command characters to display in the file logs.

deleteOnSuccess :: SwitchF p DeleteOnSuccessSwitch

If active, deletes the log file upon success.

lineTrunc :: LineTruncF p

Determines to what extent we should remove control characters from file logs.

stripControl :: ConfigPhaseF p FileLogStripControl

Strip control

type FileLoggingArgs = FileLoggingP 'ConfigPhaseArgs

type FileLoggingToml = FileLoggingP 'ConfigPhaseToml

type FileLoggingMerged = FileLoggingP 'ConfigPhaseMerged

type FileLoggingEnv = FileLoggingP 'ConfigPhaseEnv

data DeleteOnSuccessSwitch

Switch for deleting the log file upon success.

Constructors

DeleteOnSuccessOff

DeleteOnSuccessOn

mergeFileLogging :: (HasCallStack, MonadTerminal m) => FileLoggingArgs -> Maybe FileLoggingToml -> m (Maybe FileLoggingMerged)

Merges args and toml configs.

withFileLoggingEnv :: (HasCallStack, MonadFileWriter m, MonadHandleWriter m, MonadPathReader m, MonadPathWriter m, MonadSTM m, MonadTerminal m) => Maybe FileLoggingMerged -> (Maybe FileLoggingEnv -> m a) -> m a

Given merged FileLogging config, constructs a FileLoggingEnv and calls the continuation.