Shrun.Configuration.Data.Truncation

Documentation

data TruncRegion

The different regions to apply truncation rules.

Constructors

TruncCommandName
Apply truncation rules to commands/key names.

TruncLine
Apply truncation rules to command log entire lines.

newtype Truncation (a :: TruncRegion)

The maximum number of command characters to display in the logs.

Constructors

MkTruncation

Fields

unTruncation :: Natural

parseTruncation :: forall m (r :: TruncRegion). MonadFail m => m Natural -> m (Truncation r)

data LineTruncation

Determines command log line truncation behavior. We need a separate type from Truncation to add a third option, to detect the terminal size automatically.

Constructors

Undetected (Truncation 'TruncLine)

Detected

parseLineTruncation :: (Alternative m, MonadFail m) => m Natural -> m Text -> m LineTruncation

Misc

decodeCommandNameTrunc :: Decoder (Maybe (Truncation 'TruncCommandName))

decodeLineTrunc :: Decoder (Maybe LineTruncation)

configToLineTrunc :: (HasCallStack, MonadTerminal m) => WithDisabled LineTruncation -> m (Maybe (Truncation 'TruncLine))

Maps line trunc config to actual value.