Shrun.Configuration.Data.WithDisabled

Documentation

data WithDisabled a

Like Maybe but adds an extra constructor representing a "disabled" state. The idea is that both CLI Args and Toml and have optional fields, but the CLI can also be "disabled", which overrides everything.

The semigroup is similar to Maybe's:

Identity: Without

With is left-biased.

(Without, Disabled) forms a normal submonoid, in particular:

  Disabled <> _ === Disabled === _ <> Disabled

Constructors

With a
The field.

Without
Missing.

Disabled
Disabled.

Construction

fromMaybe :: Maybe a -> WithDisabled a

Nothing -> Without, Just -> With.

fromBool :: Bool -> WithDisabled ()

Elimination

toMaybe :: WithDisabled a -> Maybe a

With -> Just, o/w -> Nothing.

toBool :: WithDisabled a -> Bool

fromWithDisabled :: a -> WithDisabled a -> a

Eliminates WithDisabled.

fromDefault :: Default a => WithDisabled a -> a

Eliminates WithDisabled via its Default instance.

Misc

(<>?) :: WithDisabled a -> Maybe a -> WithDisabled a infixr 6

l <>? r lifts Maybe r into a WithDisabled per fromMaybe then runs the Semigroup.

(<>?.) :: Default a => WithDisabled a -> Maybe a -> a infixr 6

Like (<>?) except we extract a result via fromDefault.

(<>??) :: WithDisabled a -> Maybe a -> Maybe a infixr 6

Like (<>?) except we extract a Maybe via toMaybe.

Optics

_With :: Prism' (WithDisabled a) a

_Without :: Prism' (WithDisabled a) ()

_Disabled :: Prism' (WithDisabled a) ()