Shrun.Configuration.Env.Types

Description

Provides types and typeclasses for our environment.

"HasX" style typeclasses

class HasCommands env where

The commands themselves.

Methods

getCommands :: env -> NESeq CommandP1

getCompletedCommands :: env -> TVar (Seq CommandP1)

prependCompletedCommand :: (HasCallStack, HasCommands env, MonadReader env m, MonadSTM m) => CommandP1 -> m ()

Prepends a completed command.

class HasCommandLogging env where

Methods

getCommandLogging :: env -> CommandLoggingEnv

class HasCommonLogging env where

Methods

getCommonLogging :: env -> CommonLoggingEnv

class HasConsoleLogging env r where

Methods

getConsoleLogging :: env -> Tuple2 ConsoleLoggingEnv (TBQueue (LogRegion r))

class HasFileLogging env where

Methods

getFileLogging :: env -> Maybe FileLoggingEnv

class HasTimeout env where

Timeout, if any.

Methods

getTimeout :: env -> Maybe Timeout

class HasInit env where

Init, if any.

Methods

getInit :: env -> Maybe Text

class HasAnyError env where

Methods

getAnyError :: env -> TVar Bool

Retrieves the anyError flag.

setAnyErrorTrue :: (HasAnyError env, HasCallStack, MonadReader env m, MonadSTM m) => m ()

Set anyError to True.

class HasNotifyConfig env where

Class for retrieving the notify config.

Methods

getNotifyConfig :: env -> Maybe NotifyEnv

Retrieves the notify config.

Types

data Env r

The main Env type used by Shrun.

Constructors

MkEnv

Fields

config :: CoreConfigP 'ConfigPhaseEnv

completedCommands :: TVar (Seq CommandP1)

Holds a sequence of commands that have completed. Used so we can determine which commands have not completed if we time out.

consoleLogQueue :: ~(TBQueue (LogRegion r))

Console log queue.

anyError :: TVar Bool

Holds the anyError flag, signaling if any command exited with an error.

commands :: NESeq CommandP1

Holds notification environment. | Commands