Shrun.Configuration.Toml

Description

Provides the TomlConfig type.

Since: 0.5

Documentation

data TomlConfig

TomlConfig refers to TomlConfiguration we retrieve from the toml TomlConfig file.

Since: 0.5

Constructors

MkTomlConfig

Fields

timeout :: !(Maybe Timeout)

Timeout.

Since: 0.5

fileLogging :: !(Maybe FilePathDefault)

Optional file logging. If enabled, holds the path to the file and the log queue.

Since: 0.5

fileLogStripControl :: !(Maybe StripControl)

stripControl for file logging.

Since: 0.5

cmdLogging :: !(Maybe CmdLogging)

Whether to log commands.

Since: 0.5

cmdDisplay :: !(Maybe CmdDisplay)

Whether to display the command (key) names or the commands themselves.

Since: 0.5

cmdNameTrunc :: !(Maybe (Truncation 'TCmdName))

The max number of command characters to display in the logs.

Since: 0.5

cmdLineTrunc :: !(Maybe LineTruncation)

The max number of line characters to display in the logs.

Since: 0.5

stripControl :: !(Maybe StripControl)

Determines to what extent we should remove control characters from logs.

Since: 0.3

disableLogging :: !(Maybe Bool)

Overarching option for logging. If it is false then all logging is disabled.

Since: 0.5

legend :: !(Maybe (List KeyVal))

Legend text containing command aliases.

Since: 0.5

argsToTomlConfig :: Getter Args TomlConfig

Since: 0.5