Shrun.Logging.Types

Description

This module provides the Log type and associated functions.

Basic Types

data Log

Captures the relevant information concerning a specific log (i.e. command, text, level, and mode).

Constructors

MkLog

Fields

cmd :: Maybe CommandP1

Optional command that produced this log.

msg :: UnlinedText

The Text for a given log.

lvl :: LogLevel

The LogLevel for a given log.

mode :: LogMode

The LogMode for a given log.

data LogMode

Determines the logging behavior. This option only affects console logs; File logs are unaffected.

Constructors

LogModeSet
Sets the logging region to this log.

LogModeAppend
Appends the log to the logging region.

LogModeFinish
Closes the logging region, finishing with the log.

data LogLevel

Determines the logging level.

Constructors

LevelCommand

LevelFinished

LevelTimer

LevelSuccess

LevelWarn

LevelError

LevelFatal

Console Logs

data ConsoleLog

ConsoleLog is a textual representation of a given log after it has been formatted.

data LogRegion r

ConsoleLog with possible region.

Constructors

LogRegion LogMode r ConsoleLog
Log with region.

LogNoRegion ConsoleLog
Log without region.

File Logs

data FileLog

FileLog is a textual representation of a given log after it has been formatted. No coloring is included, but we include the prefix (e.g. Warn) along with a timestamp.