Shrun.Utils

Description

Provides utilities.

Since: 0.1

Text Utils

breakStripPoint :: Refined NonEmpty Text -> Text -> Tuple2 Text Text

Wrapper for Text's breakOn that differs in two ways:

Total, since we restrict the needle to NonEmpty (Text throws a pure exception here).

If the needle is found within the haystack, we do not include it in the second part of the pair.

Examples

Expand

>>> let equals = $$(R.refineTH @R.NonEmpty @Text "=")

>>> -- Data.Text
>>> T.breakOn "=" "HEY=LISTEN"
("HEY","=LISTEN")

>>> -- Shrun.Utils.Text
>>> breakStripPoint equals "HEY=LISTEN"
("HEY","LISTEN")

Other examples:

>>> breakStripPoint equals "HEYLISTEN"
("HEYLISTEN","")

>>> breakStripPoint equals "=HEYLISTEN"
("","HEYLISTEN")

>>> breakStripPoint equals "HEYLISTEN="
("HEYLISTEN","")

>>> breakStripPoint equals "HEY==LISTEN"
("HEY","=LISTEN")

Since: 0.1

decodeUtf8Lenient :: ByteString -> Text

Decodes a ByteString to UTF-8 in lenient mode.

Since: 0.1

splitOn :: HasCallStack => Refined NonEmpty Text -> Text -> NonEmptySeq Text

Wrapper around Text's splitOn. This should be total, as splitOn is partial exactly when the first parameter is empty, which we reject. Unfortunately we have to perform a partial pattern match on the result since splitOn returns a list, even though the result should always be NonEmptySeq. Hence HasCallStack.

Examples

Expand

>>> splitOn $$(R.refineTH ",,") "hey,,listen"
"hey" :|^ fromList ["listen"]

>>> splitOn $$(R.refineTH ",,") "heylisten"
"heylisten" :|^ fromList []

>>> splitOn $$(R.refineTH ",") ""
"" :|^ fromList []

Since: 0.1

truncateIfNeeded :: Natural -> Text -> Text

For Natural \(n\) and Text \(t = t_0 t_1 \ldots t_m\), truncates \(t\) if \(m > n\). In this case, \(t\) is truncated to \(n - 3\), and an ellipsis ( \(\ldots\) ) is appended. We are left with a string with length exactly \(n\):

\[ t_0 t_1 \ldots t_{n-3} \text{...} \quad \text{-- 3 literal } `\text{.' chars appended} \]

Examples

Expand

>>> truncateIfNeeded 7 "hi"
"hi"

>>> truncateIfNeeded 10 "This is 21 chars long"
"This is..."

Since: 0.1

stripControlAll :: Text -> Text

Strips all control chars, including ansi escape sequences. Leading and trailing whitespace is also stripped.

Examples

Expand

>>> stripControlAll "foo\ESC[0;3Abar \n baz"
"foobar  baz"

Since: 0.5

stripControlSmart :: Text -> Text

Strips control chars, including most ansi escape sequences. Leading and trailing whitespace is also stripped. We leave behind SGR ansi escape sequences e.g. text coloring. See https://en.wikipedia.org/wiki/ANSI_escape_code#SGR_(Select_Graphic_Rendition)_parameters.

Examples

Expand

>>> stripControlSmart "foo\ESC[0;3Abar \n baz"
"foobar  baz"

>>> stripControlSmart "foo\ESC[0;3mbar \n baz"
"foo\ESC[0;3mbar  baz"

Since: 0.5

Timing Utils

diffTime :: TimeSpec -> TimeSpec -> Natural

For given \(x, y\), returns the absolute difference \(|x - y|\) in seconds.

Examples

Expand

>>> :{
  let t1 = TimeSpec 5 0
      -- 20 s + 1 billion ns = 21 s
      t2 = TimeSpec 20 1_000_000_000
  in diffTime t1 t2
:}
16

Since: 0.1

foldMap1 :: (Foldable f, Semigroup s) => (a -> s) -> a -> f a -> s

Relaxes foldMap's Monoid constraint to Semigroup. Requires a starting value. This will have to do until semigroupoids' Foldable1 is in base.

Examples

Expand

>>> foldMap1 @List Sum 0 [1..4]
Sum {getSum = 10}

>>> -- Silly, but demonstrates usage i.e. with non-monoid NonEmpty.
>>> foldMap1 @List (:| []) 1 [2,3,4]
1 :| [2,3,4]

Since: 0.1

Refined Utils

_Refined :: Getter (Refined p a) a

Since: 0.5